requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1

# UI 
django-admin-volt==1.0.10

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0

# psycopg2-binary
# mysqlclient

django-crum
django-filter
django-simple-history